feat(vessels): clinking glasses
Issue #5585

Holding a vessel for liquids in hand, targeting a hand, and clicking on a mob that also has a vessel will allow you to clink glasses with them. Or with it. Yes, you can clink glasses with cyborgs and dionae, too. It's a feature.

SFX source: https://freesound.org/people/elricadavis/sounds/764600/ (licensed under CC0)